refactor: share wallet UI helpers in bdk_demo
